# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "HelloFresh Recipes API",
        "description": "HelloFresh Recipes API - By Juliano Petronetto",
        "contact": {
            "name": "Juliano Petronetto",
            "url": "http://petronetto.com.br"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get one user by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The recieved user",
                        "schema": {
                            "description": "Get one user by id",
                            "properties": {
                                "data": {
                                    "description": "Get one user by id",
                                    "type": "object",
                                    "$ref": "#/definitions/UserApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/definitions/Error"
                    },
                    "404": {
                        "$ref": "#/definitions/Error"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Update one user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payload",
                        "in": "body",
                        "description": "User payload",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id": {
                                    "description": "Update one user",
                                    "type": "integer",
                                    "example": 10
                                },
                                "username": {
                                    "description": "Update one user",
                                    "type": "string",
                                    "example": "jhondoe"
                                },
                                "email": {
                                    "description": "Update one user",
                                    "type": "string",
                                    "example": "john@doe.com"
                                },
                                "first_name": {
                                    "description": "Update one user",
                                    "type": "string",
                                    "example": "Jhon"
                                },
                                "last_name": {
                                    "description": "Update one user",
                                    "type": "string",
                                    "example": "Doe"
                                },
                                "password": {
                                    "description": "Update one user",
                                    "type": "string",
                                    "example": "secret"
                                },
                                "password_confirmation": {
                                    "description": "Update one user",
                                    "type": "string",
                                    "example": "secret"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated user",
                        "schema": {
                            "description": "Update one user",
                            "properties": {
                                "data": {
                                    "description": "Update one user",
                                    "type": "object",
                                    "$ref": "#/definitions/UserApiResponse"
                                }
                            }
                        },
                        "headers": {
                            "X-Powered-By": {
                                "description": ";)",
                                "type": "string"
                            },
                            "X-Response-Time": {
                                "description": "282.263ms",
                                "type": "string"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Delete one user",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "An empty response"
                    },
                    "401": {
                        "$ref": "#/definitions/Error"
                    },
                    "404": {
                        "$ref": "#/definitions/Error"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "Update one user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payload",
                        "in": "body",
                        "description": "User payload",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id": {
                                    "description": "Update one user",
                                    "type": "integer",
                                    "example": 10
                                },
                                "username": {
                                    "description": "Update one user",
                                    "type": "string",
                                    "example": "jhondoe"
                                },
                                "email": {
                                    "description": "Update one user",
                                    "type": "string",
                                    "example": "john@doe.com"
                                },
                                "first_name": {
                                    "description": "Update one user",
                                    "type": "string",
                                    "example": "Jhon"
                                },
                                "last_name": {
                                    "description": "Update one user",
                                    "type": "string",
                                    "example": "Doe"
                                },
                                "password": {
                                    "description": "Update one user",
                                    "type": "string",
                                    "example": "secret"
                                },
                                "password_confirmation": {
                                    "description": "Update one user",
                                    "type": "string",
                                    "example": "secret"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated user",
                        "schema": {
                            "description": "Update one user",
                            "properties": {
                                "data": {
                                    "description": "Update one user",
                                    "type": "object",
                                    "$ref": "#/definitions/UserApiResponse"
                                }
                            }
                        },
                        "headers": {
                            "X-Powered-By": {
                                "description": ";)",
                                "type": "string"
                            },
                            "X-Response-Time": {
                                "description": "282.263ms",
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/definitions/Error"
                    },
                    "404": {
                        "$ref": "#/definitions/Error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Error": {
            "description": "",
            "properties": {
                "type": {
                    "type": "string",
                    "example": "Exception"
                },
                "message": {
                    "type": "string",
                    "example": "Something went wrong"
                },
                "code": {
                    "type": "string",
                    "example": 400
                }
            },
            "type": "object"
        },
...
        "UserApiResponse": {
            "properties": {
                "id": {
                    "description": "API Response for Users",
                    "type": "integer",
                    "example": 10
                },
                "username": {
                    "description": "API Response for Users",
                    "type": "string",
                    "example": "jhondoe"
                },
                "email": {
                    "description": "API Response for Users",
                    "type": "string",
                    "example": "john@doe.com"
                },
                "first_name": {
                    "description": "API Response for Users",
                    "type": "string",
                    "example": "Jhon"
                },
                "last_name": {
                    "description": "API Response for Users",
                    "type": "string",
                    "example": "Doe"
                }
            },
            "type": "object"
        }
    },
    "securityDefinitions": {
        "Authorization": {
            "type": "apiKey",
            "description": "The JWT token. Get in /api/v1/auth",
            "name": "Authorization",
            "in": "header"
        }
    }
}